// prodis-prodiswcl/src/app/modules/prodisapi/model/ruolo.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Ruolo { 
    amministratore?: boolean;
    codiceFiscale?: string;
    consulenteRespo?: boolean;
    delegatoRespo?: boolean;
    denominazioneAzienda?: string;
    id?: number;
    legaleRappresentante?: boolean;
    listaCasiUso?: Array<string>;
    monitoraggio?: boolean;
    operatoreProvinciale?: boolean;
    personaAutorizzata?: boolean;
    personaAutorizzataScuola?: boolean;
    personaCaricaAziendale?: boolean;
    ruolo?: string;
    codMinSoggettoAbilitato?: string;
    idSoggetti?: number;
    cfUtente?: string;
}